Update icon with more centered version. Fix longstanding &detrung bug
Since it's inception, &detrungify and variants would make the detrung'd image "lean" to the right. Turns out this can be mitigated *somewhat* by adjusting an off-by-one error.

A TODO on this is to allow optional manual adjustments to this, for the funnies.

+++ core/image.py
@@ -70,7 +70,7 @@ def detrungify(source, updown=False):
             # - dividing the width of the image by 2*i units (so that we have 2*i evenly sized regions to pull from)
             # - multiplying that by j so that we get the j-th region
             # - subtracting one from the whole thing because of off-by-one errors
-            region_num = j * (2 * i * unit) // w
+            region_num = int((j * (2 * i * unit) // w) +1)
 
             # finally fill in the pixel.
             dest.putpixel(
